fix(community): refactor community header to avoid layout shift
